+++ src/Query/LazyQuery.php
@@ -31,17 +31,6 @@ final class LazyQuery implements QueryInterface
         return $this->getLoadedQuery()($runtime, $rootNode);
     }
 
-    /**
-     * @return bool
-     * @deprecated
-     */
-    public function isDefinite(): bool
-    {
-        return $this
-            ->getProperties()
-            ->isDefinite();
-    }
-
     public function getProperties(): QueryPropertiesInterface
     {
         return $this