+++ src/Query/QueryInterface.php
@@ -12,11 +12,5 @@ interface QueryInterface
 
     public function __invoke(RuntimeInterface $runtime, NodeValueInterface $rootNode): ValueListInterface;
 
-    /**
-     * @return bool
-     * @deprecated
-     */
-    public function isDefinite(): bool;
-
     public function getProperties(): QueryPropertiesInterface;
 }